Melhorado a tabela, adicionado permissões para os usuarios e adicionado validação para CRUD de user e WorkSchedule

+++ app/Traits/Table/WithDelete.php
@@ -16,6 +16,7 @@ trait WithDelete
 
     public function delete()
     {
+        $this->authorize($this->model->getName() . ':delete');
         $this->model->find($this->deleteId)->delete();
         $this->deleteList = [];
         $this->resetPage();
